Add docs page (#8)
- Automatically generates documentation and publish to GitHub pages on
every successful build action to the main branch
- This action will only update the docs if all tests and build actions
pass

Since the code base can grow significantly in size, I thought this
action would be necessary so we can easily access the documentation from
GitHub Pages.

This PR doesn't require an in-depth review. Docs page generation is
already tested and the output is published to
https://nocontrolco.github.io/elso.

We can later modify the Doxyfile to suite our needs.